Bug 650833: Fix crash on 32-bit Windows
Declare int_part as double on cogl/cogl-texture-2d.c to avoid overflow
in _cogl_texture_2d_wrap_coords on Windows 32-bit.  This may be a potential
problem on other 32-bit systems too.  This change also runs fine on x64
Windows, too.

+++ cogl/cogl-texture-2d.c
@@ -65,7 +65,7 @@ static void
 _cogl_texture_2d_wrap_coords (float t_1, float t_2,
                               float *out_t_1, float *out_t_2)
 {
-  float int_part;
+  double int_part;
 
   /* Wrap t_1 and t_2 to the range [0,1] */
